oauthoidc_client/requirements.txt
python-keycloak
flask==2.1.2
flask-oidc==1.4.0
itsdangerous==2.0.0
gunicorn